feat: implement SocketManager for managing socket connections and communication
